<==== Mercurial data ====>
Branch: feature_3334_time_sync

+++ common/src/core/datapacket/media_data_packet.cpp
@@ -331,7 +331,7 @@ inline bool sse4_attribute metadataIsEmpty_sse41(__m128i* src)
 inline bool metadataIsEmpty_cpu(const char* data)
 {
     const quint32* curPtr = (const quint32*) data;
-    for (int i = 0; i < MD_WIDTH*MD_HEIGHT/sizeof(quint32)/CHAR_BIT; ++i)
+    for (size_t i = 0; i < MD_WIDTH*MD_HEIGHT/sizeof(quint32)/CHAR_BIT; ++i)
     {
         if (*curPtr++)
             return false;
